requirements.txt
torch
torchvision
opencv-python
grpcio
grpcio-tools
protobuf
kafka-python
